Handle vertical layouts like those found in manhwa
Because manhwa are often to long to convert to a single image, a PDF
document is used instead. To ensure that the PDF displays properly
without any gaps between the pages, the input images are first resized
to have uniform width and then split, appended, and extended to acheive
uniform height.