Implement read command
Without arguments the read command starts a local server and attempts to
open the user's default web browser to a page displaying a list of the
available manga along with their direction, source, and progress.
Opening the browser may be disabled by the --no-open flag.

Passing the title of a manga will open it in a simple viewer with a
continuous vertical scrolling layout or a page-by-page horizontal layout
as appropriate. As the user navigates, a bookmark of their current
location will be saved. The --last flag may be used to continue reading
the most recently accessed manga from the bookmarked location.

The --port flag can be used to change the port the server listens on.